Adding Diffing Capabilities and Additional Curation Facilities to Match
Added the FineDiff package to compare pages and their suggested edits.
Added a load of routes and methods and tests around the curation
facilities to allow approving and saving over of pages and stuff. Also
rejecting nasty edits. Also some other stuff? And a page about
contributing in case anyone is insane enough to be interested in that.