Connect Using HTTPS Not HTTP
Our SSH config uses HTTP over port 8000. This is probably not a good
idea. In dev-standards port 8000 has been removed because why would you
connect via port HTTP when you can use HTTPS.

So this updates the links to use https:// and port 8443.

There is an issue where the "http" part of the URL is in the database. I
will create a seporate pull request for this as there are multiple
solutions for this problem.